==> content/notebooks/01_context.ipynb <==
# 1. Pregunta orientadora

**🤔 ¿Cuál es el impacto de las energías renovables en Colombia para el 2025?**

#🖥️ Análisis de los datos

- El conjunto de datos proporciona estadísticas mensuales de electricidad de la Agencia Internacional de Energía (AIE) para múltiples países y territorios en todo el mundo.
- Incluye información sobre la generación de electricidad mensual desde 2010 hasta 2022.
- La producción de energía se mide en gigavatios-hora (GWh) y abarca una variedad de productos energéticos, incluyendo hidroeléctrica, eólica, solar, geotérmica, nuclear, combustibles fósiles y otros.
- Estos datos pueden ser útiles para analizar las tendencias de consumo y producción de electricidad a nivel mundial y regional, así como para evaluar el impacto ambiental de la industria eléctrica.

## 🕸️ Fuentes
- https://www.kaggle.com/code/joeportilla/producci-n-mensual-de-energ-a-en-colombia-eda/notebook
- https://api.iea.org/mes/list/ - Monthly Electricity Statistics

## 💾 Categorias

- COUNTRY: Name of the country
- CODE_TIME: A code that represents the month and year (e.g., JAN2010 for January 2010)
- TIME: The month and year in a more human-readable format (e.g., January 2010)
- YEAR: The year of the data point
- MONTH: The month of the data point as a number (1-12)
- MONTH_NAME: The month of the data point as a string (e.g., January)
- PRODUCT: The type of energy product (e.g., Hydro, Wind, Solar)
- VALUE: The amount of electricity generated in gigawatt-hours (GWh)
- DISPLAY_ORDER: The order in which the products should be displayed
- yearToDate: The amount of electricity generated for the current year up to the current month in GWh
- previousYearToDate: The amount of electricity generated for the previous year up to the current month in GWh
- share: The share of the product in the total electricity generation for the country in decimal format

# 🎯 Objetivo general
Aplicar los metodos de análisis unidimensiosal, multidimensional, supervisados y no superevisados para obtener el impacto que puedan tener las energias renovables en la generación de energía en Colombia.

## Objetivo especifico
- Los ejercicios realizados se centra en explorar y visualizar el conjunto de datos.

- Realizar un análisis exploratorio de datos para comprender los datos.

# Rango de la data
Enero de 2010 hasta Diciembre de 2022
