==> content/notebooks/03_clear_data.ipynb <==
# import library
import pandas as pd
import os

# set default directory
os.chdir('/home/lbertel/code/talento_tech/content')
!ls

# load raw file
df_data_all_country = pd.read_csv('data/raw/data.csv')

# print all data
df_data_all_country.info()

# select row with COUNTRY is equal Colombia
df_data_colombia = df_data_all_country[df_data_all_country['COUNTRY'] == 'Colombia']

# print first record of Colombia
df_data_colombia.head(10)

# size dataframe
len(df_data_colombia)

# show columns after delete
print(df_data_colombia.columns)

# delete columns
df_temp = df_data_colombia.drop(['COUNTRY', 'CODE_TIME', 'TIME', 'MONTH_NAME', 'DISPLAY_ORDER', 'yearToDate', 'previousYearToDate', 'share'], axis=1)

# show columns before delete
print(df_temp.columns)
print('')
print('---------------------------------')
print('')

df_temp.info()

# delete row with null values
len(df_temp)
df_without_null = df_temp.dropna(how='any')

# size dataframe without null
len(df_without_null)

# eliminate row with PRODUCT containt Total
filter_total = df_without_null[~df_without_null['PRODUCT'].str.contains('Total', case=False, na=False)]

len(filter_total)

# save the new dataframe
filter_total.to_csv('data/processed/colombia_data.csv', index=False)

# information dataframe
filter_total.info()

# dimension of dataframe
filter_total.shape

# missing values
filter_total.isnull().sum(axis=0)

# missing values
filter_total.isnull().sum(axis=0)
