==> content/notebooks/04_unidimensional.ipynb <==
# import library
%matplotlib inline
import os
import pandas as pd

# set default directory
os.chdir('/home/lbertel/code/talento_tech/content')

# load data
colombia_df = pd.read_csv('data/processed/colombia_data.csv')
colombia_df.head(10)

## Visión general del datase

## Análisis unidimensional
